INBOmd.bib: solve issue with series/booktitle Title Case *
Title Case phrases must be enclosed by {} if they are to
be retained by a CSL processor or by the Better BibTeX
plugin of Zotero.
Assuming sentence case for these fields is the normal
behaviour and is in general recommended
(see ).
Converting to title/sentence case for an outputted field can
be imposed by the style, but conversion from title case to
sentence case is error-prone due to proper nouns.

Hence, an alternative would also be to always convert to
Title Case within the style.