INBOmd.bib: use sentence case for webpage title *
Otherwise, for consistent behaviour with Title Case, such values must be surrounded by {Curly Braces} in order to keep as Title Case. Not doing so leads to inconsistent behaviour between languages and to loosing Title Case. Here it has been preferred to drop title case, given that it was not relevant for the example.

For more information, see https://inbomd-examples.netlify.app/citation_style/nl/3-bibcsl.html#voor-pandoc-gebruikt-in-r-markdown-en-bij-import-in-zotero-indien-gebruik-van-title-case-strings .